Refactor: 알림 아이템 -> 알림 리스트까지를 features로 분류
