--- features/notification-item/container/NotificationItemCont.tsx
@@ -1,13 +1,0 @@
-import NotificationItemPres from '../presentational/NotificationItemPres';
-import { NotificationItemType } from '../types';
-
-interface Props {
-  item: NotificationItemType;
-  deleteMode: boolean;
-  selected: boolean;
-  onToggleSelect: (id: number) => void;
-}
-
-export default function NotificationItemCont(props: Props) {
-  return <NotificationItemPres {...props} />;
-}
