--- features/notification-item/index.ts
@@ -1,1 +1,0 @@
-export { default as NotificationItemCont } from './container/NotificationItemCont';
