+++ features/notification-list/index.ts
@@ -1,0 +1,1 @@
+export { default as NotificationListCont } from './container/NotificationListCont';
